+++ cfgov/v1/models/learn_page.py
@@ -11,8 +11,7 @@ from wagtail.wagtailadmin.edit_handlers import TabbedInterface, ObjectList, \
     StreamFieldPanel, FieldPanel, FieldRowPanel, MultiFieldPanel, InlinePanel
 from wagtail.wagtailimages.edit_handlers import ImageChooserPanel
 
-from . import molecules
-from . import organisms
+from ..atomic_elements import molecules, organisms
 from .base import CFGOVPage, CFGOVPageManager
